requirements.txt
numpy
pandas
astropy
scipy
tk
Pillow
matplotlib
astroalign
